requirements.txt
python-telegram-bot>=20.0
requests
sentence-transformers
faiss-cpu
numpy
langdetect
deep_translator
apscheduler